GCP/3. base project/requirements-func.txt
flask==2.3.3
google-cloud-firestore
numpy
requests
google-cloud-pubsub
gunicorn
wtforms
flask-restful